Use VS code debugging icons in debugger
