+++ packages/devtools_app/lib/src/app_size/app_size_screen.dart
@@ -15,10 +15,10 @@ import '../config_specific/drag_and_drop/drag_and_drop.dart';
 import '../config_specific/server/server.dart' as server;
 import '../config_specific/url/url.dart';
 import '../notifications.dart';
-import '../octicons.dart';
 import '../screen.dart';
 import '../split.dart';
 import '../theme.dart';
+import '../ui/icons.dart';
 import '../utils.dart';
 import 'app_size_controller.dart';
 import 'app_size_table.dart';
