+++ packages/devtools_app/lib/src/inspector/inspector_screen.dart
@@ -13,11 +13,11 @@ import '../blocking_action_mixin.dart';
 import '../common_widgets.dart';
 import '../connected_app.dart';
 import '../globals.dart';
-import '../octicons.dart';
 import '../screen.dart';
 import '../service_extensions.dart' as extensions;
 import '../split.dart';
 import '../theme.dart';
+import '../ui/icons.dart';
 import '../ui/service_extension_widgets.dart';
 import 'inspector_controller.dart';
 import 'inspector_screen_details_tab.dart';
