--- packages/devtools_app/lib/src/octicons.dart
@@ -1,12 +1,0 @@
-import 'package:flutter/material.dart';
-
-class Octicons {
-  static const IconData bug = IconData(61714, fontFamily: 'Octicons');
-  static const IconData info = IconData(61778, fontFamily: 'Octicons');
-  static const IconData deviceMobile = IconData(61739, fontFamily: 'Octicons');
-  static const IconData fileZip = IconData(61757, fontFamily: 'Octicons');
-  static const IconData clippy = IconData(61724, fontFamily: 'Octicons');
-  static const IconData package = IconData(61812, fontFamily: 'Octicons');
-  static const IconData dashboard = IconData(61733, fontFamily: 'Octicons');
-  static const IconData pulse = IconData(61823, fontFamily: 'Octicons');
-}
