+++ packages/devtools_app/lib/src/performance/performance_screen.dart
@@ -17,11 +17,11 @@ import '../config_specific/import_export/import_export.dart';
 import '../dialogs.dart';
 import '../globals.dart';
 import '../notifications.dart';
-import '../octicons.dart';
 import '../screen.dart';
 import '../service_extensions.dart';
 import '../split.dart';
 import '../theme.dart';
+import '../ui/icons.dart';
 import '../ui/service_extension_widgets.dart';
 import '../ui/utils.dart';
 import '../ui/vm_flag_widgets.dart';
