+++ packages/devtools_app/lib/src/profiler/profiler_screen.dart
@@ -16,9 +16,9 @@ import '../common_widgets.dart';
 import '../config_specific/import_export/import_export.dart';
 import '../globals.dart';
 import '../notifications.dart';
-import '../octicons.dart';
 import '../screen.dart';
 import '../theme.dart';
+import '../ui/icons.dart';
 import '../ui/vm_flag_widgets.dart';
 import 'cpu_profile_controller.dart';
 import 'cpu_profile_model.dart';
