+++ packages/devtools_app/lib/src/ui/icons.dart
@@ -251,3 +251,23 @@ Image createImageIcon(String url, {double size = defaultIconSize}) {
     width: size,
   );
 }
+
+class Octicons {
+  static const IconData bug = IconData(61714, fontFamily: 'Octicons');
+  static const IconData info = IconData(61778, fontFamily: 'Octicons');
+  static const IconData deviceMobile = IconData(61739, fontFamily: 'Octicons');
+  static const IconData fileZip = IconData(61757, fontFamily: 'Octicons');
+  static const IconData clippy = IconData(61724, fontFamily: 'Octicons');
+  static const IconData package = IconData(61812, fontFamily: 'Octicons');
+  static const IconData dashboard = IconData(61733, fontFamily: 'Octicons');
+  static const IconData pulse = IconData(61823, fontFamily: 'Octicons');
+}
+
+// Unicode values can be looked up from https://github.com/microsoft/vscode-codicons/blob/master/src/template/mapping.json.
+class Codicons {
+  static const IconData resume = IconData(60111, fontFamily: 'Codicon');
+  static const IconData pause = IconData(60113, fontFamily: 'Codicon');
+  static const IconData stepIn = IconData(60116, fontFamily: 'Codicon');
+  static const IconData stepOut = IconData(60117, fontFamily: 'Codicon');
+  static const IconData stepOver = IconData(60118, fontFamily: 'Codicon');
+}
